requirements/local.txt
MySQL-python==1.2.5
# Not going to use this;
django-grappelli==2.6.3
Django-Select2==4.2.2
South==1.0.2
aldryn-blog==0.4.6
aldryn-common==0.0.6
django-appdata==0.1.3
django-hvad==1.0.0
django-taggit==0.11.2
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-link==1.5
djangocms-picture==0.1
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-twitter==0.0.1
djangocms-video==0.1
